Migrate Transforms and ImageKey to dc_bundle
The Transforms types are basically re-exports of euclid types, so dc_bundle will need to temporarily depend on euclid
