+++ crates/dc_bundle/Cargo.toml
@@ -12,5 +12,9 @@ prost.workspace = true
 thiserror.workspace = true
 serde.workspace = true
 
+# Temporary dependencies during the transition to protobuf structures
+euclid.workspace = true
+
+
 [build-dependencies]
 prost-build.workspace = true
