+++ crates/dc_bundle/src/legacy_definition/element.rs
@@ -17,6 +17,7 @@
 pub mod color;
 pub mod font;
 pub mod geometry;
+pub mod image;
 pub mod node;
 pub mod path;
 pub mod variable;
