+++ crates/figma_import/src/transform_flexbox.rs
@@ -19,8 +19,7 @@ use crate::toolkit_font_style::FontWeight;
 use crate::toolkit_layout_style::Overflow;
 
 use crate::toolkit_style::{
-    Background, GridLayoutType, GridSpan, LayoutTransform, MeterData, ShadowBox, StyledTextRun,
-    TextStyle, ViewStyle,
+    Background, GridLayoutType, GridSpan, MeterData, ShadowBox, StyledTextRun, TextStyle, ViewStyle,
 };
 
 use crate::figma_schema;
@@ -46,6 +45,7 @@ use dc_bundle::legacy_definition::layout::positioning::{
 use dc_bundle::legacy_definition::modifier::blend::BlendMode;
 use dc_bundle::legacy_definition::modifier::filter::FilterOp;
 use dc_bundle::legacy_definition::modifier::text::{TextAlign, TextAlignVertical, TextOverflow};
+use dc_bundle::legacy_definition::modifier::transform::LayoutTransform;
 use log::error;
 use unicode_segmentation::UnicodeSegmentation;
 //::{Taffy, Dimension, JustifyContent, Size, AvailableSpace, FlexDirection};
