+++ crates/figma_import/src/vector_schema.rs
@@ -15,9 +15,10 @@
 use std::convert::TryFrom;
 
 use dc_bundle::legacy_definition::modifier::blend::BlendMode;
+use dc_bundle::legacy_definition::modifier::transform::AffineTransform;
 use serde::{Deserialize, Serialize};
 
-use crate::toolkit_style::{AffineTransform, Background, BoxShadow, Stroke};
+use crate::toolkit_style::{Background, BoxShadow, Stroke};
 
 // This is a simple vector representation that should reproduce Figma or
 // another design tool's rendering. It simplifies the renderer's task by
